requirements.txt
fastapi
uvicorn
pytest
pytest-cov
pylint
python-dotenv
mysql-connector-python
sqlalchemy
sqlalchemy-utils
pika
aio_pika
mock
jose
python-jose
passlib
python-multipart
pyjwt